✨ refactor: update imports to centralize intervention types
- Refactor imports in intervention-related components to source types from '@/types/intervention' for improved maintainability.
- Remove redundant type definitions from interventionTypeService to streamline the codebase.
- Add eslint-disable comments in useEffect hooks to address warnings related to dependencies.
