+++ src/app/dashboard/(registre-securite)/interventions/intervention-types/_components/InterventionTypeCard.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import { InterventionType } from '@/services/interventionTypeService';
+import { InterventionType } from '@/types/intervention';
 import {
   Delete as DeleteIcon,
   Edit as EditIcon
